backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
psycopg[binary]
python-dotenv==1.0.0
Flask-CORS==5.0.0
PyJWT==2.8.0
google-genai
Flask-Migrate==4.0.5
Flask-Limiter>=3.5.0
gunicorn>=21.2.0
